[Refactor] #119 - MVI 통합
