+++ Projects/Home/HomeFeature/Sources/Home/Store/HomeStore.swift
@@ -15,14 +15,30 @@ enum HomeIntent {
     case onAppear
     case onErrorToastDisappear
     case onToastDisappear
-    case onDeleteInterestConcert
+
+    // Interest concert related
+    case onDelete
     case _fetchUserInterestConcertResult(Result<Concert?, Error>)
+    case _fetchScheduleListResult(Result<ConcertScheduleList, Error>)
+    case _fetchMainSetlistResult(Result<Setlist, Error>)
+    case _fetchSetlistSongListResult(Result<SetlistSongList, Error>)
+    case _deleteInterestConcertResult(Result<Void, Error>)
 }
 
 struct HomeState {
     var mode: HomeMode = .noInterestedConcert
     var toastMessage: String = ""
     var errorMessage: String = ""
+
+    // Interest concert substate
+    var scheduleList: ConcertScheduleList = []
+    var setlist: Setlist?
+    var songList: SetlistSongList = []
+
+    var interestConcert: Concert? {
+        if case .hasInterestedConcert(let concert) = mode { return concert }
+        return nil
+    }
 }
 
 final class HomeStore: ObservableObject {
@@ -31,27 +47,69 @@ final class HomeStore: ObservableObject {
     @Injected private var repository: HomeRepository
 
 
-
     @MainActor
     func send(_ intent: HomeIntent) {
         switch intent {
         case .onAppear:
             performFetchUserInterestedConcert()
-            
+            // If we already have an interest concert loaded, also fetch related details
+            if let concert = state.interestConcert {
+                performFetchScheduleList(concertID: concert.id)
+                performFetchMainSetlist(concertID: concert.id)
+            }
+
         case .onErrorToastDisappear:
             state.errorMessage = ""
         
         case .onToastDisappear:
             state.toastMessage = ""
 
-        case .onDeleteInterestConcert:
-            state.mode = .noInterestedConcert
-            state.toastMessage = "관심 콘서트가 삭제되었어요"
+        case .onDelete:
+            performDeleteInterestConcert()
 
         case ._fetchUserInterestConcertResult(let result):
             switch result {
             case .success(let concert):
                 state.mode = concert.map { .hasInterestedConcert($0) } ?? .noInterestedConcert
+                // Fetch related data when we have a concert
+                if let concert = concert {
+                    performFetchScheduleList(concertID: concert.id)
+                    performFetchMainSetlist(concertID: concert.id)
+                }
+            case .failure(let error):
+                state.errorMessage = error.localizedDescription
+            }
+
+        case ._fetchScheduleListResult(let result):
+            switch result {
+            case .success(let schedules):
+                state.scheduleList = schedules
+            case .failure(let error):
+                state.errorMessage = error.localizedDescription
+            }
+
+        case ._fetchMainSetlistResult(let result):
+            switch result {
+            case .success(let setlist):
+                state.setlist = setlist
+            case .failure(let error):
+                state.errorMessage = error.localizedDescription
+            }
+
+        case ._fetchSetlistSongListResult(let result):
+            switch result {
+            case .success(let songs):
+                state.songList = songs
+            case .failure(let error):
+                state.errorMessage = error.localizedDescription
+            }
+
+        case ._deleteInterestConcertResult(let result):
+            switch result {
+            case .success:
+                // update UI
+                state.mode = .noInterestedConcert
+                state.toastMessage = "관심 콘서트가 삭제되었어요"
             case .failure(let error):
                 state.errorMessage = error.localizedDescription
             }
@@ -69,6 +127,54 @@ private extension HomeStore {
                 await send(._fetchUserInterestConcertResult(.success(result)))
             } catch {
                 await send(._fetchUserInterestConcertResult(.failure(error)))
+            }
+        }
+    }
+
+    func performFetchScheduleList(concertID: Int) {
+        Task {
+            do {
+                let schedules = try await repository.fetchScheduleList(for: concertID)
+                await send(._fetchScheduleListResult(.success(schedules)))
+            } catch {
+                await send(._fetchScheduleListResult(.failure(error)))
+            }
+        }
+    }
+
+    func performFetchMainSetlist(concertID: Int) {
+        Task {
+            do {
+                let setlist = try await repository.fetchMainSetlist(for: concertID)
+                let songList = try await repository.fetchSongList(for: setlist.id)
+                await send(._fetchMainSetlistResult(.success(setlist)))
+                await send(._fetchSetlistSongListResult(.success(songList)))
+            } catch HomeError.noResponse {
+                return
+            } catch {
+                await send(._fetchMainSetlistResult(.failure(error)))
+            }
+        }
+    }
+
+    func performFetchSetlistSongList(setlistID: Int) {
+        Task {
+            do {
+                let songList = try await repository.fetchSongList(for: setlistID)
+                await send(._fetchSetlistSongListResult(.success(songList)))
+            } catch {
+                await send(._fetchSetlistSongListResult(.failure(error)))
+            }
+        }
+    }
+    
+    func performDeleteInterestConcert() {
+        Task {
+            do {
+                try await repository.deleteInterestedConcert()
+                await send(._deleteInterestConcertResult(.success(())))
+            } catch {
+                await send(._deleteInterestConcertResult(.failure(error)))
             }
         }
     }
